requirements.txt
fastapi==0.110
uvicorn==0.29.0
colorama==0.4.6
google-cloud-storage==2.10.0
google-cloud==0.34.0
google-cloud-secret-manager==2.16.4
requests
pytest==8.1.1
pytest-cov==4.1.0
sqlalchemy==2.0.29
redis==5.0.3
pymongo==4.6.2
python-dotenv==1.0.1
chalice==1.31.2
celery[redis]
psycopg2-binary
debugpy==1.8.0
grpcio==1.59.0
grpcio-tools==1.59.0
fuzzywuzzy==0.18.0
sentry_sdk==2.7.1
watchdog==4.0.1